tests #16 - Adicionando mais testes para a socialworkerservice
